Split caching into general case and specific external package cases
